refactor: extract reified toolchain probe
